+++ programs/staking-program/src/state/nft_stake.rs
@@ -8,9 +8,9 @@ pub struct NftStake {
     pub user: Pubkey,
     pub nft_mint: Pubkey,
     pub is_active: bool,
-    pub staked_on: u64,
-    pub unstaked_on: u64,
-    pub last_claimed: u64,
+    pub staked_on: i64,
+    pub unstaked_on: i64,
+    pub last_claimed: i64,
     pub reward_amount: u64,
 
     pub bump: u8,
@@ -27,15 +27,15 @@ impl NftStake {
 
         let days_elapsed = Self::days_elapsed(unix_timestamp);
         let token_emssion = 100_000;
-        return token_emssion * days_elapsed;
+        return (token_emssion * days_elapsed) as u64;
     }
 
-    fn days_elapsed(unix_timestamp: u64) -> u64 {
+    fn days_elapsed(unix_timestamp: i64) -> i64 {
         let now = SystemTime::now()
             .duration_since(SystemTime::UNIX_EPOCH)
             .unwrap();
-        let then = Duration::from_secs(unix_timestamp);
+        let then = Duration::from_secs(unix_timestamp as u64);
         let elapsed = now - then;
-        elapsed.as_secs() / 86_400 // 86,400 seconds in a day
+        (elapsed.as_secs() / 86_400) as i64 // 86,400 seconds in a day
     }
 }
